Added java-doc and comments to encryption class.
